requirements.txt
# Dependências do Projeto - Análise de Processos Acessórias

# Core
requests>=2.31.0
python-dotenv>=1.0.0

# Data Processing (versões compatíveis sem compilação)
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Date/Time
python-dateutil>=2.8.2

# Logging
colorlog>=6.8.0

# Utilities
tqdm>=4.66.1

# ============ BACKEND API ============
# FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# CORS
python-multipart>=0.0.6

# ============ WHATSAPP GESTOR ============
# WhatsApp API (opcional - escolher uma)
# twilio>=8.10.0  # Se usar Twilio (pago)
# requests já instalado acima para Meta Cloud API ou Evolution API

# Segurança e Validação
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# HTTP Client
httpx>=0.25.0

# Rate Limiting (opcional)
# slowapi>=0.1.9
